Handeling Bad Request to the api

+++ app/Http/Controllers/UserInvitesController.php
@@ -10,25 +10,45 @@ class UserInvitesController extends Controller
 {
     public static function received($id)
     {
-        $result = User::find($id)->invitesReceived()->get();
-        return response()->json($result, 200);
+        $user= User::find($id);
+        if($user){
+            $result = $user->invitesReceived()->get();
+            return response()->json($result, 200);
+        }
+        else{
+            return response()->json(["Message"=>"User not found"],404);
+        }
     }
     public static function sent($id)
     {
-        $result = User::find($id)->invitesSent()->get();
-        return response()->json($result);
+        $user= User::find($id);
+        if($user){
+            $result = $user->invitesSent()->get();
+            return response()->json($result, 200);
+        }
+        else{
+            return response()->json(["Message"=>"User not found"],404);
+        }
     }
     public static function accept($userId, $inviteId)
     {
         $invite = Invite::find($inviteId);
-        $invite->project()->first()->collaborators()->save(User::find($userId));
-        $invite->delete();
-        return response()->json(201);
+        if ($invite) {
+            $invite->project()->first()->collaborators()->save(User::find($userId));
+            $invite->delete();
+            return response()->json(["Message" => "Accepted Invite"], 201);
+        } else {
+            return response()->json(["Message" => "Invite is no longer available"], 404);
+        }
     }
     public static function decline($userId, $inviteId)
     {
         $invite = Invite::find($inviteId);
-        $invite->delete();
-        return response()->json();
+        if ($invite) {
+            $invite->delete();
+            return response()->json(["Message" => "Declined Invite"], 200);
+        } else {
+            return response()->json(["Message" => "Invite is no longer available"], 404);
+        }
     }
 }
